academic_agent/requirements.txt
# 核心依赖
requests>=2.28.0
pyyaml>=6.0

# 数据处理
pandas>=1.5.0
numpy>=1.23.0

# 网络分析
networkx>=2.8.0

# HTTP服务
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0

# 缓存（可选）
redis>=4.5.0

# 数据导出
openpyxl>=3.1.0
xlsxwriter>=3.0.0

# LLM集成
openai>=1.0.0
anthropic>=0.18.0

# 测试
pytest>=7.2.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# 工具
python-dotenv>=1.0.0
tqdm>=4.65.0
